reorganize folders for helpers and small component
