+++ client/src/components/errorPage/error.scss
