+++ client/src/components/errorPage/errorPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Back from '../../buttons/backButton'; // ? Jack - I feel like importing components with a different name is not a good thing, can we look into this?
+import Back from '../buttons/backButton'; // ? Jack - I feel like importing components with a different name is not a good thing, can we look into this?
 import './error.scss';
 
 function Error (): JSX.Element {
