+++ client/src/components/jobsPage/jobListing/jobDetails/jobDetails.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { decode } from 'he';
 
 import { useUserContext } from '../../../contexts/user';
-import { numberFormatter } from '../../../helpers';
+import { numberFormatter } from '../../../utils';
 import Anchor from '../../../buttons/anchor';
 import Button from '../../../buttons/primaryButton';
 import Map from './map/map';
