+++ client/src/components/jobsPage/sidebar/utils/index.ts
