+++ client/src/components/jobsPage/sidebar/utils/userDetails.test.jsx
