+++ client/src/components/welcome/utils/index.ts
