+++ client/src/components/welcome/utils/welcomeMessage.tsx
