+++ client/src/components/welcome/utils/welcomeText.tsx
